+++ R/curve_rev.R
@@ -51,7 +51,7 @@ curve_rev <- function(point,
 
   # Produce Consonance / Surprisal Functions --------------------------------
 
-  #Moved outside the "c" conditional; otherwise skipped in "l"
+  # Moved outside the "c" conditional; otherwise skipped in "l"
   if (is.numeric(point) != TRUE) {
     stop("Error: 'point' must be a numeric value")
   }
@@ -60,37 +60,36 @@ curve_rev <- function(point,
     stop("Error: 'measure' must be a string such as 'default' or 'ratio'")
   }
 
-  if (is.null(se) && (is.null(LL) & is.null(UL))){
+  if (is.null(se) && (is.null(LL) & is.null(UL))) {
     stop("se or UL & LL must be entered")
   }
 
   if (is.null(se)) {
-
-  if (is.numeric(LL) != TRUE) {
-    stop("Error: 'LL' must be a numeric value")
-  }
-  if (is.numeric(UL) != TRUE) {
-    stop("Error: 'UL' must be a numeric value")
-  }
+    if (is.numeric(LL) != TRUE) {
+      stop("Error: 'LL' must be a numeric value")
+    }
+    if (is.numeric(UL) != TRUE) {
+      stop("Error: 'UL' must be a numeric value")
+    }
   }
 
   intrvls <- (1:steps) / steps
   z <- qnorm(1 - intrvls / 2)
-  conf.level_two = conf.level + (1 - conf.level)/2
-  conf_range = qnorm(conf.level_two)
+  conf.level_two <- conf.level + (1 - conf.level) / 2
+  conf_range <- qnorm(conf.level_two)
 
 
-  if (!is.null(se) ) {
-    LL = point - conf_range*se
-    UL = point + conf_range*se
+  if (!is.null(se)) {
+    LL <- point - conf_range * se
+    UL <- point + conf_range * se
   }
 
   if (is.null(se) && measure == "default") {
-    se <- (UL - LL) / (2*conf_range)
+    se <- (UL - LL) / (2 * conf_range)
   }
 
   if (is.null(se) && measure == "ratio") {
-    se <- log(UL / LL) / (2*conf_range)
+    se <- log(UL / LL) / (2 * conf_range)
     print(se)
   }
 
@@ -98,13 +97,13 @@ curve_rev <- function(point,
 
   if (type == "c") {
 
-    #intrvls <- (1:steps) / steps
-    #z <- qnorm(1 - intrvls / 2)
-    #conf.level_two = (1-conf.level)/2
-    #conf_range = qnorm(conf.level_two)
+    # intrvls <- (1:steps) / steps
+    # z <- qnorm(1 - intrvls / 2)
+    # conf.level_two = (1-conf.level)/2
+    # conf_range = qnorm(conf.level_two)
 
     if (measure == "default") {
-      #se <- (UL - LL) / (2*conf_range)
+      # se <- (UL - LL) / (2*conf_range)
       LL <- pbmclapply(z, FUN = function(i) point + (i * se), mc.cores = getOption("mc.cores", 1L))
       UL <- pbmclapply(z, FUN = function(i) point - (i * se), mc.cores = getOption("mc.cores", 1L))
       df <- data.frame(do.call(rbind, UL), do.call(rbind, LL))
@@ -113,7 +112,7 @@ curve_rev <- function(point,
     }
 
     else if (measure == "ratio") {
-      #se <- log(UL / LL) / (2*conf_range)
+      # se <- log(UL / LL) / (2*conf_range)
       logpoint <- log(point)
       logLL <- pbmclapply(z, FUN = function(i) logpoint + (i * se), mc.cores = getOption("mc.cores", 1L))
       logUL <- pbmclapply(z, FUN = function(i) logpoint - (i * se), mc.cores = getOption("mc.cores", 1L))
@@ -151,11 +150,11 @@ curve_rev <- function(point,
 
     # Produce Likelihood / Deviance Functions ---------------------------------
   } else if (type == "l") {
-    #intrvls <- (1:steps) / steps
-    #z <- qnorm(1 - intrvls / 2)
+    # intrvls <- (1:steps) / steps
+    # z <- qnorm(1 - intrvls / 2)
 
     if (measure == "default") {
-      #se <- (UL - LL) / (2*conf_range)
+      # se <- (UL - LL) / (2*conf_range)
       LL <- pbmclapply(z, FUN = function(i) point + (i * se), mc.cores = getOption("mc.cores", 1L))
       UL <- pbmclapply(z, FUN = function(i) point - (i * se), mc.cores = getOption("mc.cores", 1L))
       df <- data.frame(do.call(rbind, UL), do.call(rbind, LL))
@@ -165,7 +164,7 @@ curve_rev <- function(point,
     }
 
     else if (measure == "ratio") {
-      #se <- log(UL / LL) / (2*conf_range)
+      # se <- log(UL / LL) / (2*conf_range)
       logpoint <- log(point)
       logLL <- pbmclapply(z, FUN = function(i) logpoint + (i * se), mc.cores = getOption("mc.cores", 1L))
       logUL <- pbmclapply(z, FUN = function(i) logpoint - (i * se), mc.cores = getOption("mc.cores", 1L))
@@ -184,13 +183,13 @@ curve_rev <- function(point,
     class(df) <- c("data.frame", "concurve")
 
 
-    #se <- log(UL / LL) / (2*conf_range)
-    if ( measure == "ratio") {
-    values <- seq(from = df[1, 1], to = df[1, 2], by = 0.01)
-    zscore <- sapply(
-      values,
-      function(j) (log(j / point) / se)
-    )
+    # se <- log(UL / LL) / (2*conf_range)
+    if (measure == "ratio") {
+      values <- seq(from = df[1, 1], to = df[1, 2], by = 0.01)
+      zscore <- sapply(
+        values,
+        function(j) (log(j / point) / se)
+      )
     }
 
     if (measure == "default") {
@@ -204,9 +203,9 @@ curve_rev <- function(point,
     support <- exp((-zscore^2) / 2)
     deviancestat <- (zscore^2)
     if (measure == "ratio") {
-    likelihood <- support * (log(point))
+      likelihood <- support * (log(point))
     } else {
-      likelihood = support * point
+      likelihood <- support * point
     }
     loglikelihood <- log(likelihood)
     likfunction <- data.frame(values, likelihood, loglikelihood, support, deviancestat)